slurm.spec - reimplement sysv init script support
Add local System V init versions of systemd's scriptlet macros, and a
slurm_scriptlet macro to select the systemd or init version based on
init system in use. This builds packages that only include code for
the init system in use, and keeps the scriptlet part of the spec file
clean.

Use System V init on el6 by default. To build with System V init on
other distributions --without=systemd can be used.

The upgrade path from old package names with separate slurmctld and
slurmd packages is not completely handled. If slurmctld or slurmd is
running before the upgrade, they will be stopped rather than
restarted. The new separate slurmctld and slurmd services have to be
started manually the first time.