Modify killproc usage in init scripts
Previously daemons was only sent TERM.

Use killproc default behaviour instead, which is to first send TERM
and then after a delay if send KILL if the process is still
running. Set the delay to 60 seconds to give the daemons plenty of
time to shut down.

For SlurmDBD this change also ensures that the pid file is deleted.